requirements.txt
pywebview[gtk]
Pillow
pytesseract
